fix: code review — critical strategy bugs, state machine safety, ccxt compliance
* fix(strategy): correct Bellman-Ford cycle detection and per-leg quantity propagation

C-3 graph.py:
- Use a virtual-source initialisation so dist[v] correctly tracks shortest
  paths reachable from any node, instead of all-zero init.
- Run n-1 relaxation passes followed by a dedicated n-th detection pass;
  any further relaxation signals a node on (or downstream of) a negative
  cycle.
- Store the relaxation weight on each pred entry so the cycle walk no longer
  has to re-scan the edge list. Eliminates the type: ignore[index] hack.
- Return None cleanly when a one-way single edge cannot form a cycle.

C-2 triangular.py:
- Replace identical per-leg quantities with a propagated chain: leg1 uses
  the configured base quantity, then each subsequent leg's order quantity
  is the previous leg's output amount converted into the next pair's
  base-asset units (BUY divides by ask, SELL passes through).
- Compute the conversion in a new _compute_quantities helper that handles
  BUY/SELL output and input semantics separately.
- Add a 5-second price TTL so stale cached prices no longer feed the arb
  search.


* fix(execution): exception safety, DB state persistence, ccxt error hygiene

C-5 state.py:
- Wrap the leg sequence in try/except so any unexpected exception unwinds
  whichever legs filled, persists status FAILED if unwind also fails, and
  always returns the state machine to IDLE via finally.
- Re-raise asyncio.CancelledError so cooperative shutdown propagates.
- Replace print()-based logging with logging.getLogger(__name__).

C-6 state.py:
- Collapse _db_finish_arb into _db_update_arb_status and invoke it on the
  LEG1_FILLED and LEG2_FILLED transitions, so intermediate progress is
  visible if a leg later fails.

H-1 binance.py:
- Split the subscribe loop's blanket "except Exception" into specific
  branches: re-raise CancelledError and AuthenticationError, sleep+retry on
  NetworkError/RequestTimeout/ExchangeNotAvailable, log+raise on anything
  unexpected.

H-2 binance.py:
- Pass newClientOrderId so retried submissions are idempotent at the
  exchange, and route quantity/price through ccxt's amount_to_precision /
  price_to_precision helpers.

H-3 binance.py + exchange_interface.py:
- cancel_order now accepts (order_id, symbol) so callers must pass the
  exchange-assigned order id rather than the local UUID. ExchangeInterface
  ABC and test doubles updated to match.

H-4 binance.py:
- Track the subscribe task on the adapter and cancel/await it in close()
  so shutdown is graceful and asyncio.CancelledError propagates cleanly.


* fix(risk): notional-based exposure check and StrEnum RiskDecision

C-1 manager.py:
- Replace the apples-to-oranges base-quantity comparison with a USDT
  notional check. Each leg is valued via the most recent USDT mid price
  for its base asset; if the price is missing, the signal is rejected.
- Add update_prices() so a live price feed can refresh the cached USDT
  prices used by check().
- max_order_size and max_exposure are now both interpreted as USDT
  notional, with the old defaults in run_local.py replaced by realistic
  per-leg/total caps for the testnet.

M-4 manager.py:
- RiskDecision is now a StrEnum, and check() returns RiskDecision rather
  than a plain str.

engine.py / live/runner.py:
- ExecutionEngine.update_prices(...) and LiveRunner now extract the
  USDT-quoted tickers and push the mid prices into the risk manager on
  every on_tickers call, so the notional check has fresh inputs.


* fix(alerts): wire broadcast path and fix set.discard bug

C-4 ws_alerts.py:
- Replace the list-typed _clients with a set so the existing
  _clients.discard(client) call on a failed send actually works (lists
  have no .discard).
- broadcast() iterates over a snapshot of the set to avoid mutation
  during iteration when a client fails mid-send.

api-server routes/internal.py:
- Add /internal/alert POST endpoint that accepts a JSON payload and
  forwards it through broadcast() to every connected dashboard.

core-platform outbox/workers.py:
- Extend AlertWorker with an optional http_url destination. Either,
  both, or neither of webhook_url / http_url can be set. Failed HTTP
  posts log via logging.getLogger but never crash the loop.

run_local.py:
- Add the ALERT_API_URL env var (defaulting to
  http://localhost:8000/internal/alert) so the worker auto-fan-outs to
  the dashboard during local testnet runs.


* refactor(quality): type safety, Decimal boundary, EventBus cleanup, stale price TTL

M-1 exchange_interface.py:
- TickerCallback is now Callable[[dict[str, Any]], Awaitable[None]] so
  type checkers can verify subscribe_ticker call sites.

M-2 core/trading_pair.py:
- TradingPair switches to (frozen=True, kw_only=True) for consistency
  with the rest of the dataclass surface.

M-3 exchange/order_result.py + exchange/binance.py + execution/state.py:
- price/average/amount/filled/remaining/cost are now Decimal | None.
  Conversion happens once at the binance boundary via _to_decimal();
  the state machine no longer re-wraps result.filled in Decimal(str(...)).

M-5 market/feed.py + live/runner.py:
- Remove the duplicate runner → strategy path via the event bus. The
  runner is now the single owner of strategy invocation; the feed just
  feeds the tick queue. EventBus stays available for other event types.

M-6 execution/order.py:
- Add Order.exchange_id so callers can carry the exchange-assigned id
  alongside the local UUID.

M-7 events/bus.py:
- publish() uses asyncio.gather(..., return_exceptions=True) so a buggy
  subscriber no longer blocks the fan-out; failures are logged.

M-8 market/feed.py:
- Ticks with missing/zero timestamps are dropped during feed conversion.

M-9 core/parsers.py:
- parse_trading_pair rejects ccxt futures symbols ("BASE/QUOTE:SETTLE").

M-12 web-dashboard/src/api/alerts.ts + AlertFeed.tsx:
- Alert objects carry a stable id (`${arb_id}:${ts}`) and the JSX uses it
  as the React key instead of the array index.

M-13 api-server/app/models.py + pyproject.toml:
- Drop the duplicated ArbAttempt definition. The api-server now imports
  ArbAttempt/OrderRecord from atlas.db.models via the new
  `atlas-trading` workspace dependency.

M-14 api-server/app/main.py:
- CORS origins are read from the ATLAS_CORS_ORIGINS env var (default:
  http://localhost:5173) instead of being hard-coded to "*".


* fix(db): add indexes for arb_attempts and raw_ticks; misc quality fixes

H-5 alembic/versions/0002_add_hot_path_indexes.py:
- New migration that creates ix_arb_attempts_created_at and
  ix_arb_attempts_status (for the trades/positions API queries) and
  ix_raw_ticks_symbol_ts (for tick replay/analytics).

L-5 outbox/queue.py:
- OutboxEntry.payload is now dict[str, Any] so outbox events can carry
  numeric/timestamp fields without lying about their type.

L-8 web-dashboard/src/api/trades.ts:
- fetchTrades now raises on non-2xx responses instead of silently passing
  whatever JSON the server returned.

L-9 web-dashboard/src/components/TradeTable.tsx:
- Track mounted state in a ref and gate setState calls so late
  fetch responses can't mutate an unmounted component.

(L-2 print → logging in state.py was already addressed in the state-
machine bundle; run_local.py CLI prints are intentional and untouched.)


* test(arb): edge case tests + fix crossed book, negative price, partial fill

- Reject crossed book (bid > ask) and invalid prices in graph._build_edges
  and strategy._update_prices
- Propagate actual fill quantity to subsequent legs in ArbitrageStateMachine
- Add 39 edge case tests covering price validation, TTL, partial fills,
  concurrent signals, unwind paths, and fee gap documentation


---------

+++ core-platform/atlas/exchange/order_result.py
@@ -1,4 +1,5 @@
 from dataclasses import dataclass
+from decimal import Decimal
 
 from atlas.execution.order_status import OrderStatus
 
@@ -12,12 +13,12 @@ class OrderResult:
     side: str | None = None
     timestamp: int | None = None
     datetime: str | None = None
-    price: float | None = None
-    average: float | None = None
-    amount: float | None = None
-    filled: float | None = None
-    remaining: float | None = None
-    cost: float | None = None
+    price: Decimal | None = None
+    average: Decimal | None = None
+    amount: Decimal | None = None
+    filled: Decimal | None = None
+    remaining: Decimal | None = None
+    cost: Decimal | None = None
     client_order_id: str | None = None
     time_in_force: str | None = None
     post_only: bool | None = None
